requirements.txt
RPi.GPIO
urllib3
Flask
Adafruit_DHT
elasticsearch
uwsgi
